moved timer state change to parent componenet - PomodoroTimer.js so state can also be accessed in setting view, thus preventing timerstate when user switches between timer/setting mode
